made PRCE working

+++ src/downloads.c
@@ -90,7 +90,7 @@ uint8_t has_been_downloaded(const simple_list bucket, const char *url) {
 int addToBucket(const char* identifier, NODE **head, const int maxBucketItems) {
 
 	addItem(am_strdup(identifier), head);
-	if(maxBucketItems > 0 && listCount(*head) > maxBucketItems) {
+	if(maxBucketItems > 0 && listCount(*head) > (uint32_t)maxBucketItems) {
 		dbg_printf(P_INFO, "[add_to_bucket] bucket gets too large, deleting head item...\n");
 		removeLast(*head, NULL);
 	}
